requirements.txt
openai>=1.12.0
PyMuPDF>=1.24.0
PyYAML>=6.0
pydantic>=2.0
tiktoken>=0.5.0
tenacity>=8.2.0
python-dotenv>=1.0.0
python-docx>=0.8.11
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
